refactor: dedupe review-flagged helpers and fix lint line count
readLockFileOrFail now lives once in lib/lock-read (re-exported from
commands/utils), enabledHarnesses once in lib/harness, plural once in
commands/utils, ensureDir once in lock-copy. Drops the redundant
HarnessSyncResult.linked field (always equal to synced). lint's line
counter no longer counts the phantom element after a trailing newline.
CLAUDE.md refreshed (pi harness, current lockfile schema, module map).
